Steps 2 of #8
Continulation of #8
