+++ assistant_manager/init.py
@@ -5,10 +5,11 @@ from . import assistant_chat
 from . import interface_base
 from . import a_m_threads
 from . import oai_base
+from . import runs_manager
 
 
 # Init file for the assistant_manager package.
 #
 # This file is used to initialize the assistant_manager package.
 
-__all__ = ["assistant_manager", "a_m_threads", "assistant_chat" , "assistant_tools", "autogen_assistant_manager", "autogen_assistant_tools", "interface_base", "oai_assistant", "oai_base"]
+__all__ = ["assistant_manager", "a_m_threads", "assistant_chat" , "assistant_tools", "autogen_assistant_manager", "autogen_assistant_tools", "interface_base","runs_manager", "oai_assistant", "oai_base"]
